requirements.txt
torch>=1.12.0
torchvision>=0.13.0
diffusers>=0.13.0
transformers>=4.25.1
Pillow>=9.0.0
numpy>=1.20.0
tqdm>=4.64.0
ftfy>=6.1.1
lpips>=0.1.4
matplotlib>=3.5.0